Make the domain layer not dependant on Data Layer
So, based on Domain Layer in Clean Architecture explaination :
The domain layer or the domain package is the innermost and the only independent layer among the three layers. When I say innermost, I mean that this package/layer is not accessible to the outside world and when I say that it’s independent, it’s meant that this package is not dependent on any other package or class for its working (However, its dependent on the di package and we will come to that later). Domain layer consists mainly of our app’s business logic. By business logic, I mean the model classes (POJO classes in Java or data classes in Kotlin), use cases , requests, etc,.

Thanks to : https://medium.com/@gangulysourik/a-definitive-guide-to-clean-architecture-in-android-with-mvvm-d74a0533ef2c

I have make the domain layer now fully not dependant to another layer and also i have make the Data Source now (Remote / Layer) on the data layer since the
Domain Layer only consist on app's business logic. It doesn't need to know where the data source came from.