refactor: Refactor project structure (read commit notes)
I realize that my project structure is a bit, premature.
In order to more better separation of concern, to be inline with Clean Architecture principles, i decided to separate the presentation
layer into it's own layer.

And also, with those refactor, i have to reconfigure the product flavoring configuration, which thankfully it's working.

References :
https://developer.android.com/build/build-variants#kts